Updated README.md Added source and target compatibilities in the build.gradle Moved WidgetEventFactory to the factory package
