+++ build.gradle
@@ -11,6 +11,9 @@ version '0.15.0'
 
 mainClassName = 'bomb.Main'
 
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
+
 compileJava {
     options.encoding("UTF-8")
 }
